+++ internal/server/server.go
@@ -42,7 +42,7 @@ import (
 )
 
 var (
-	Version     = "1.9.0"
+	Version     = "1.10.0"
 	connCounter uint64
 )